Extract draw preview, add draw drag
